+++ src/Announcers/FinalRankAnnouncer/FinalRank.cs
@@ -11,8 +11,24 @@ public static class FinalRankPatcher
     [HarmonyPatch(nameof(FinalRank.FlashPanel))]
     public static void FlashPanelPatch(ref GameObject panel, FinalRank __instance)
     {
-        if (panel == __instance.toAppear[__instance.i].transform.parent.GetChild(1).gameObject)
-            LogHelper.LogInfo($"object name = {__instance.toAppear[__instance.i]}");
-        LogHelper.LogInfo("FlashPanel Triggerd");
+        if (panel != __instance.toAppear[__instance.i].transform.parent.GetChild(1).gameObject) return;
+        LogHelper.LogDebug($"FinalRank.i(index) : {__instance.i}");
+        var index = __instance.i;
+        if (index == 0)
+        {
+            
+        }
+        if (index == 1)
+        {
+            
+        }
+        if (index == 2)
+        {
+            
+        }
+        if (index == 3)
+        {
+            
+        }
     }
 }