+++ src/Announcers/RankAnnouncer/StyleHUD.cs
@@ -35,38 +35,3 @@ public static class StyleHUDPatcher
         }
     }
 }
-
-
-
-/*[HarmonyPatch(typeof(StyleHUD), "UpdateMeter")]  // For D rank only, left for review silly code, why do you want to patch a Update() function? 
->>>>>>> 145108309970536fe7b8cd4ad469d6e00f945bb7
-public static class StyleHUDUpdateMeter_Patch
-{
-    private static readonly AccessTools.FieldRef<StyleHUD, float> currentMeterRef = AccessTools.FieldRefAccess<StyleHUD, float>("currentMeter");
-    private static          bool                                  previousWasZero = true;
-    static void Postfix(StyleHUD __instance)
-    {
-        float currentMeter    = GetCurrentMeter(__instance);
-        bool currentIsNonZero = __instance.rankIndex == 0 && currentMeter > 0;
-
-        if (previousWasZero && currentIsNonZero)
-        {
-            RankAnnouncer.PlaySound(0);
-        }
-
-        previousWasZero = __instance.rankIndex == 0 && currentMeter <= 0;
-    }
-
-    
-
-    private static float GetCurrentMeter(StyleHUD instance)
-    {
-        return currentMeterRef(instance);
-    }
-
-    private static float GetCurrentMeter(StyleHUD instance)
-    {
-        return Traverse.Create(instance).Field("currentMeter").GetValue<float>();
-    }
-}*/
-